update user info storage, report API integration, and misc fixes

+++ src/store/modules/user.js
@@ -1,13 +1,26 @@
 import { createSlice } from '@reduxjs/toolkit';
 import { deleteCookie } from 'utils/cookie';
 import mixpanel from 'mixpanel-browser';
-// import { setCookie, getCookie } from 'utils/cookie';
+
+const dev_user = {
+  userId: "6374fec7-65d3-40b0-a9a0-4dbec96eef75",
+  nickname: "DEV_USER",
+  email: "user001@gmail.com",
+  preference: "T",
+  profileImageEnabled: false,
+  emailAdsConsented: false,
+  agreeToTerms: false,
+  agreeToPrivacyPolicy: false,
+  firstLogin: false
+}
 
 const initialState = {
-  userInfo: null,
+  userInfo: process.env.REACT_APP_ENV === 'development' ? dev_user : null,
   isLoggedIn: null,
   sessionLoading: true,
+  policiesLimit: null
 };
+
 const userSlice = createSlice({
   name: 'user',
   initialState,
@@ -21,13 +34,21 @@ const userSlice = createSlice({
         state.isLoggedIn = true;
       }
     },
+    setPoliciesLimit: (state, action) => {
+      if (action.payload) {
+        state.policiesLimit = action.payload;
+      }
+    },
     clearUserInfo:(state) => {
       mixpanel.reset();
-      state.userInfo = null;
-      state.isLoggedIn = null;
-      deleteCookie('--user-data');
-      deleteCookie('jwt-access');
-      deleteCookie('jwt-refresh');
+      if (process.env.REACT_APP_ENV === 'production') {
+        state.userInfo = null;
+        state.isLoggedIn = null;
+        localStorage.removeItem('--policies-limit');
+        localStorage.removeItem('--user-data');
+        deleteCookie('jwt-access');
+        deleteCookie('jwt-refresh');
+      }
     },
   },
 });
@@ -37,4 +58,5 @@ export const user = (state) => state.user;
 export const { setUserInfo,
   clearUserInfo,
   setSessionLoading,
+  setPoliciesLimit
  } = userSlice.actions;
